refactor(server): move the reload lifecycle into documents/watcher
ADR-034 Unit 7a, part 2. `reloadFromDisk` and `wireFileWatcher` are a
self-contained loop -- watch, decide conflict-vs-reload, repopulate,
re-anchor -- that shared nothing with the open entries beyond the file
they were written in. With them out, document-service's last dynamic
`import("./file-opener")` becomes static: both cycles it was breaking are
gone rather than deferred.

`reloadInProgress` travels with them, and that is the part worth reading.
It is the per-document concurrent-reload guard, and `reloadDocumentFromMarkdown`
takes the same one so two clear+repopulate transactions never interleave
on one Y.Doc. Leaving it behind would make watcher.ts import back into
mcp/; exporting the Set would let a second module desynchronize it. So it
gets a named acquire/release/probe contract instead.

That turns a structural guarantee into a cross-module one, and a mutation
showed nothing observed it: a private Set in file-opener left all 785
existing specs green with the serialization simply gone. Two specs now
assert it from both sides -- file-opener refusing while the watcher holds
the guard, and the watcher's own probe seeing file-opener's acquisition.
Both go red under that mutation.

The one edge from watcher.ts back into mcp/ is annotation sanitization,
recorded as residue in the boundary inventory rather than hidden.
