+++ tests/docs/documents-boundary.test.ts
@@ -286,6 +286,8 @@ const FAN_IN = [
   // import document-service to ask a question about a map it writes itself.
   "server/mcp/document-service.ts -> server/documents/annotation-wiring.ts (value) x1",
   "server/mcp/document-service.ts -> server/documents/conflict.ts (value) x1",
+  "server/mcp/file-opener.ts -> server/documents/watcher.ts (value) x1",
+  "server/mcp/document-service.ts -> server/documents/watcher.ts (value) x1",
   "server/mcp/presence-expiry.ts -> server/documents/registry.ts (value) x1",
   "server/mcp/routes/backups.ts -> server/documents/registry.ts (value) x1",
   "server/mcp/routes/document-raw.ts -> server/documents/registry.ts (value) x1",
@@ -329,6 +331,24 @@ const FAN_OUT = [
   "server/documents/conflict.ts -> shared/origins.ts (value) x1",
   "server/documents/conflict.ts -> shared/types.ts (type) x1",
   "server/documents/conflict.ts -> shared/utils.ts (value) x1",
+  "server/documents/watcher.ts -> server/annotations/doc-hash.ts (value) x1",
+  "server/documents/watcher.ts -> server/annotations/migration-log.ts (value) x1",
+  "server/documents/watcher.ts -> server/events/queue.ts (value) x1",
+  "server/documents/watcher.ts -> server/file-io/index.ts (value) x1",
+  "server/documents/watcher.ts -> server/file-watcher.ts (value) x1",
+  // The one edge out of documents/ that points back at mcp/, and the reason
+  // this list is phrased as residue rather than as a contract: annotation
+  // sanitization has not been split out of the MCP layer yet.
+  "server/documents/watcher.ts -> server/mcp/annotations.ts (value) x1",
+  "server/documents/watcher.ts -> server/notifications.ts (value) x1",
+  "server/documents/watcher.ts -> server/positions.ts (value) x1",
+  "server/documents/watcher.ts -> server/yjs/provider.ts (value) x1",
+  "server/documents/watcher.ts -> shared/constants.ts (value) x1",
+  "server/documents/watcher.ts -> shared/origins.ts (value) x1",
+  "server/documents/watcher.ts -> shared/positions/types.ts (value) x1",
+  "server/documents/watcher.ts -> shared/snapshot.ts (value) x1",
+  "server/documents/watcher.ts -> shared/types.ts (type) x1",
+  "server/documents/watcher.ts -> shared/utils.ts (value) x1",
   "server/documents/open.ts -> server/mcp/file-opener.ts (value) x1",
   "server/documents/registry.ts -> server/yjs/provider.ts (value) x1",
   "server/documents/registry.ts -> shared/constants.ts (value) x1",
@@ -545,7 +565,6 @@ describe("runtime export surfaces", () => {
         "reloadDocumentFromMarkdown",
         "resolveExternalConflict",
         "restoreDocumentFromBackup",
-        "wireFileWatcher",
       ].sort(),
     );
   });
